Completed (I think) the float to double conversion. 1) All calculations and prior <float> NOT specific to the GUI are doubles EXCEPT for the    productivity value from the database, as for whatever reason the    default is <float> for that, so I kept that <float>. 2) All GUI values remain as <float>, which required making it rain with    the (float) explicit casts where underlying <double> based analysis    results are used in display functions. .NET 8 performs a rounding function to    perform the explicit cast, it does not simply shed bits and bitshift.
build.sh completes successfully and the Calculator runs. Beginning
testing now.

+++ Yafc/Widgets/ImmediateWidgets.cs
@@ -91,8 +91,8 @@ public static class ImmediateWidgets {
 
     public static bool BuildFloatInput(this ImGui gui, DisplayAmount amount, TextBoxDisplayStyle displayStyle, bool setInitialFocus = false) {
         if (gui.BuildTextInput(DataUtils.FormatAmount(amount.Value, amount.Unit), out string newText, null, displayStyle, true, setInitialFocus)
-            && DataUtils.TryParseAmount(newText, out float newValue, amount.Unit)) {
-            amount.Value = newValue;
+            && DataUtils.TryParseAmount(newText, out double newValue, amount.Unit)) {
+            amount.Value = (float)newValue;
             return true;
         }
